Look for Sensor instead of WeatherSensor
